requirements.txt
llama-index>=0.10.0
llama-index-llms-google-genai>=0.2.0
llama-index-embeddings-google-genai>=0.1.0
google-generativeai>=0.7.0
duckdb>=0.9.2
duckdb-engine>=0.10.0
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0

llama-index-embeddings-huggingface>=0.2.0
sentence-transformers>=2.6.1
google-genai>=0.6.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
